Feedbacks on PR and added kill
Could not use setup because continious did not get updated accordingly to the props. So if you got a swiper that had one picture, the continuous flag would be set false and stay false since setup would not use any new props.

Kill had to be used because there were mulitple listeners which created problems for swiping.