Create LICENSE
Hi Dmitri,

At our company, we use your module. We are very happy with it, thank you very much.
We are currently working on a project that only allows us to use external dependencies that include a license.
I took the opportunity to create one for you in this PR.

Appreciate very much if you can take a look at this, we would love to continue to use your package. Please let me know if there is anything I could do for you.